LangGraph-Adaptive-RAG/pyproject.toml
[project]
name = "langgraph-adaptive-rag"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "faiss-cpu>=1.11.0",
    "langchain>=0.3.25",
    "langchain-community>=0.3.24",
    "langchain-google-genai>=2.1.4",
    "langchain-ollama>=0.3.3",
    "langchain-tavily>=0.1.6",
    "langchain-teddynote>=0.3.45",
    "langgraph>=0.4.5",
    "pdfplumber>=0.11.6",
    "pytest>=8.3.5",
    "python-dotenv>=1.1.0",
]
